feat(widgets): allow negative bar and run anchors
Bar and TextRun exposed unsigned offsets, though the commands they
encode carry i16. Neither widget reached left of or above its anchor
cell, while Polyline points and Icon offsets already did.

The kit's own callers derive their offsets from the saturating
sixteenths helpers. A new helper narrows those offsets to the signed
anchor, because a plain cast wraps a saturated value into a small
negative one.

+++ stoat/src/render/editor.rs
@@ -298,7 +298,7 @@ pub(crate) fn render_editor_with_overlay(
                     };
                     Bar {
                         x: 0,
-                        y: row_offset * 16,
+                        y: (row_offset * 16) as i16,
                         width: 6,
                         height: 16,
                         color,
