+++ stoat/src/render/hover.rs
@@ -154,7 +154,7 @@ pub(crate) fn render_hover(stoat: &mut Stoat, buf: &mut Buffer, scene: &mut ApcS
                         if seg_text.is_empty() {
                             continue;
                         }
-                        let col = (seg_start as u16 * TEXT_SCALE_POPUP + 8) / 16;
+                        let col = ((seg_start as u16 * TEXT_SCALE_POPUP + 8) / 16) as i16;
                         let bg = if selected {
                             selection.map_or(run_bg, |(_, _, rgb)| rgb)
                         } else {
@@ -430,7 +430,7 @@ pub(crate) fn render_hover_page(
         let line = truncate_line(line, region_width as usize);
         let mut chars_before = 0u16;
         for (text, style) in &line {
-            let col = (chars_before * TEXT_SCALE_POPUP + 8) / 16;
+            let col = ((chars_before * TEXT_SCALE_POPUP + 8) / 16) as i16;
             let color = crate::render::review::style_rgb(style.fg).unwrap_or(modal_fg);
             TextRun {
                 col,
